Refactor CharmProject class into its own submodule
The CharmProject module is large and needs its own space.  It also needs
a bunch of tests but that will happen shortly.
